Add shell default handling to workflows
