+++ modules/ghlint-api/src/main/kotlin/net/twisterrob/ghlint/model/Defaults.kt
@@ -1,0 +1,15 @@
+package net.twisterrob.ghlint.model
+
+public interface Defaults {
+
+	public val run: Run?
+
+	public interface Run {
+
+		public val shell: String?
+
+		public companion object
+	}
+
+	public companion object
+}
