+++ modules/ghlint-api/src/main/kotlin/net/twisterrob/ghlint/model/Workflow.kt
@@ -10,6 +10,7 @@ public interface Workflow : Component {
 	public val env: Map<String, String>?
 	public val jobs: Map<String, Job>
 	public val permissions: Map<String, String>?
+	public val defaults: Defaults?
 
 	public companion object
 }
